refactor(rel-import): header, headermenu, uishell updated
Carbon imports changed to relative proxy path
