+++ src/components/UIShell/index.js
@@ -1,0 +1,7 @@
+export {
+  Content,
+  Switcher,
+  SwitcherItem,
+  SwitcherDivider,
+  SkipToContent,
+} from 'carbon-components-react';
